refactor :: enum global로 이주
