+++ src/main/kotlin/com/dsm/fanflow/domain/log/domain/Log.kt
@@ -1,8 +1,8 @@
 package com.dsm.fanflow.domain.log.domain
 
-import com.dsm.fanflow.domain.log.domain.group.Group
 import com.dsm.fanflow.domain.user.domain.User
-import com.dsm.fanflow.global.BaseIdEntity
+import com.dsm.fanflow.global.domain.BaseIdEntity
+import com.dsm.fanflow.global.domain.enum.Group
 import javax.persistence.Column
 import javax.persistence.Entity
 import javax.persistence.EnumType
@@ -12,7 +12,7 @@ import javax.persistence.ManyToOne
 
 @Entity
 class Log(
-    override val id: Long , // 부모 클래스에서 상속받은 id 필드
+    override val id: Long,
 
     var title: String,
     var content: String,
