+++ src/main/kotlin/com/dsm/fanflow/domain/log/domain/repository/LogRepository.kt
@@ -1,8 +1,7 @@
 package com.dsm.fanflow.domain.log.domain.repository
 
 import com.dsm.fanflow.domain.log.domain.Log
-import com.dsm.fanflow.domain.log.domain.group.Group
-import java.util.UUID
+import com.dsm.fanflow.global.domain.enum.Group
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
