+++ src/main/kotlin/com/dsm/fanflow/domain/log/facade/LogFacade.kt
@@ -1,8 +1,8 @@
 package com.dsm.fanflow.domain.log.facade
 
 import com.dsm.fanflow.domain.log.domain.Log
-import com.dsm.fanflow.domain.log.domain.group.Group
 import com.dsm.fanflow.domain.log.domain.repository.LogRepository
+import com.dsm.fanflow.global.domain.enum.Group
 import org.springframework.stereotype.Component
 
 @Component
