--- src/main/kotlin/com/dsm/fanflow/domain/user/domain/role/Role.kt
@@ -1,5 +1,0 @@
-package com.dsm.fanflow.domain.user.domain.role
-
-enum class Role {
-    USER, ADMIN
-}
