+++ src/main/kotlin/com/dsm/fanflow/global/domain/enum/Group.kt
@@ -1,4 +1,4 @@
-package com.dsm.fanflow.domain.log.domain.group
+package com.dsm.fanflow.global.domain.enum
 
 enum class Group {
     AB6IX, SVT, TXT, RIZE, ITZY, NCT_127, NCT_DREAM, LUCY, MONSTAX, IVE, LE_SSERAFIM, NEWJEANS, BTS, TBZ, I_DLE, AESPA, ETC
