+++ src/main/kotlin/com/dsm/fanflow/global/domain/enum/Role.kt
@@ -1,0 +1,5 @@
+package com.dsm.fanflow.global.domain.enum
+
+enum class Role {
+    USER, ADMIN
+}
